Max students input styling

+++ src/app/groups/group-details/group-details.component.ts
@@ -2,7 +2,13 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Observable } from 'rxjs';
-import { DataService, Group, Student, Teacher, Week } from 'src/app/data.service';
+import {
+  DataService,
+  Group,
+  Student,
+  Teacher,
+  Week,
+} from 'src/app/data.service';
 import { EnrollStudentDialogComponent } from './enroll-student-dialog/enroll-student-dialog.component';
 import { EnrolledStudentsService } from './enrolled-students.service';
 import { TimeDialogComponent } from './time-dialog/time-dialog.component';
@@ -51,8 +57,10 @@ export class GroupDetailsComponent implements OnInit {
   }
 
   deleteStudent(student: Student): void {
-    this.dataService.deleteEnrolledStudent(this.group, student);
-    this.studentsService.refresh(this.group);
+    if (!this.group.closed) {
+      this.dataService.deleteEnrolledStudent(this.group, student);
+      this.studentsService.refresh(this.group);
+    }
   }
 
   editTime(): void {
